Split database file to separate files for each class
