+++ barbucket/db_initializer.py
@@ -6,15 +6,7 @@ import logging
 from .mediator import Mediator
 
 
-class DbNotInitializedError(Exception):
-    """Custom exception for connecting to a non-present database."""
-
-    def __init__(self, message: str) -> None:
-        self.message = message
-        super().__init__(message)
-
-
-class DatabaseInitializer():
+class DbInitializer():
     def __init__(self, mediator: Mediator = None) -> None:
         self.mediator = mediator
         conf_path = self.mediator.notify(
@@ -31,10 +23,10 @@ class DatabaseInitializer():
         new_name = self._DB_PATH.parent / f"database_archived_{timestamp}.db"
         try:
             self._DB_PATH.rename(new_name)
-        except FileNotFoundError:
-            logging.exception(
-                "Database not archived, because no databse exists.")
-            # raise FileNotFoundError from e
+        except FileNotFoundError as e:
+            logging.exception("Database not archived, because no databse "
+                              "exists.")
+            raise FileNotFoundError from e
         else:
             logging.info(f"Database archived as: {new_name}")
 
@@ -156,38 +148,3 @@ class DatabaseInitializer():
             logging.info("Database created. Finished initialization.")
         else:
             logging.info("Database already exists. Finished initialization.")
-
-
-class DatabaseConnector():
-    """Handles all non-specific database operations."""
-
-    def __init__(self, mediator: Mediator = None) -> None:
-        self.mediator = mediator
-        conf_path = mediator.notify(
-            "get_config_value_single",
-            {'section': "database", 'option': "db_location"})
-        self._DB_PATH = Path.home() / Path(conf_path)
-
-    def connect(self) -> sqlite3.Connection:
-        """
-        Provides a 'connection'-object for the database.
-        Because in sqlite, PRAGMA orders need to be given to the connection object.
-        """
-
-        if not self._DB_PATH.is_file():
-            raise DbNotInitializedError(
-                message="Database does not exist. Call 'init_database()' before connecting.")
-        # otherwise connect() would create an empty file
-
-        conn = sqlite3.connect(self._DB_PATH)
-        conn.execute("""
-            PRAGMA foreign_keys = 1;
-        """)
-        # On SQLite, PRAGMA commands are only valid for the existing connection,
-        # so this needs to be part of every db connection.
-        return conn
-
-    def disconnect(self, conn: sqlite3.Connection) -> None:
-        """Disconnects the connection to the database."""
-
-        conn.close()
